### 2025-01-17 #7688 @editorialbot
[PRE REVIEW]: AniSOAP: Machine Learning Representations for Coarse-grained and Non-spherical Systems
**Submitting author:** @rosecers (<a href="http://orcid.org/0000-0003-4515-3441">Rose K. Cersonsky</a>)
**Repository:** https://github.com/cersonsky-lab/AniSOAP/
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
