### 2025-01-17 648eed5
adjusted acknowledgements and moved it before references
