+++ paper/paper.md
@@ -9,21 +9,27 @@ authors:
   - name: Arthur Yan Lin
     orcid: 0000-0002-7665-3767
     affiliation: 1
+    equal-contrib: true
   - name: Lucas Ortengren
-    orcid: 0000-0000-0000-0000
+    orcid: 0009-0002-8899-7513
     affiliation: 1
+    equal-contrib: false
   - name: Seonwoo Hwang
     orcid: 0000-0000-0000-0000
     affiliation: 1
+    equal-contrib: false
   - name: Yong-Cheol Cho
     orcid: 0009-0001-6038-6764
     affiliation: 1
+    equal-contrib: false
   - name: Jigyasa Nigam
     orcid: 0000-0001-6857-4332
     affiliation: 2
+    equal-contrib: false
   - name: Rose K. Cersonsky
     orcid: 0000-0003-4515-3441
     affiliation: 1
+    corresponding: true
 affiliations:
  - name: Department of Chemical and Biological Engineering, University of Wisconsin-Madison, USA
    index: 1
@@ -53,23 +59,15 @@ While CG has made enormous strides in accelerating simulation and analysis, it's
 By incorporating anisotropy, AniSOAP extends the advances of atomistic machine learning to coarse-grained and mesoscale systems. With the flexibility of machine learning algorithms and anchored by the concepts underlying the feature-rich SOAP representation, AniSOAP provides a promising conduit for providing data-driven insights to many chemical systems at different length and time-scales.
 
 # What is an AniSOAP vector?
-An AniSOAP vector is a series of numbers that represents a system of ellipsoidal particles, which in the most general case, consists of $\geq 1$ particles with different positions, shapes, orientations, and species identifications. Essentially, this is very similar to other point-cloud representations used in atomistic machine learning, like SOAP[@bartok_representing_2013] or ACE[@drautz_atomic_2019], except for the fact that AniSOAP vectors incorporate additional geometric information. 
-
-The AniSOAP vector captures $n$-body correlations between different particles, which means that it encodes information about a central particle and its interactions with $(n-1)$ of its neighbors. In the majority of cases, we take $n=3$, which we call "the power spectrum", which encodes information about each particle, and the distance and subtended angle from the central particle to each possible pair of two neighboring particles within a cutoff radius. Ocassionally, we take $n=2$, called "the radial spectrum" which only captures all pairwise interactions up to a cutoff radius. 
-
-The radial spectra is built by expanding local atom density fields in terms of radial basis functions and spherical harmonics -- this construction ensures that the values are translationally, rotationally, and permutationally invariant. To obtain higher body order, we take higher order correlations of these radial spectra. These entries are stored in a multidimensional array using the `TensorMap` format in the package `metatensor`[@guillaume_fraux_metatensor_2024]; this multidimensional array is usually flattened into a 1-dimensional representation to be used in ML algorithms. These vectors are information-rich, and can be used in various machine learning models. Traditionally, the power spectrum is used as an input to shallower machine learning models, like gaussian processes or kernel/linear regresssion, to learn system energies and forces. The AniSOAP vector can also be used as an order parameter to describe phase transitions. Specific use-cases of AniSOAP can be seen in our paper here[@lin_expanding_2024].
+An AniSOAP vector is a series of numbers that represents a system of ellipsoidal particles, which in the most general case, consists of $\geq 1$ particles with different positions, shapes, orientations, and species identifications. Essentially, this is very similar to other point-cloud, many-body representations used in atomistic machine learning, like SOAP[@bartok_representing_2013] or ACE[@drautz_atomic_2019], except for the fact that AniSOAP vectors incorporate additional geometric information. Specific use-cases of AniSOAP can be seen in our paper here[@lin_expanding_2024].
 
 # What does the AniSOAP package do to take a configuration and construct the representation?
[…]
