### 2025-01-17 a9a976f
Update .readthedocs.yaml
rust 1.70 --> rust 1.77
