### 2025-01-17 d7b2aa7
Update .readthedocs.yaml
1.77 -> 1.78 (1.77 was not valid option for rtd)
