+++ .readthedocs.yaml
@@ -4,7 +4,7 @@ build:
   os: "ubuntu-22.04"
   tools:
     python: "3.10"
-    rust: "1.77"
+    rust: "1.78"
 
 python:
   install:
