+++ paper/paper.md
@@ -15,7 +15,6 @@ authors:
     affiliation: 1
     equal-contrib: false
   - name: Seonwoo Hwang
-    orcid: 0000-0000-0000-0000
     affiliation: 1
     equal-contrib: false
   - name: Yong-Cheol Cho
