### 2025-01-21 9a621d0
removed equal-contrib (it's not working properly)
