### 2025-01-22 #7688 @rosecers
@crvernon the formatting issues with the document should be fixed, as well as the paper length.
