### 2025-02-13 #7688 @arthur-lin1027
@mbobra We'd like to suggest [Samuel Tovey](https://reviewers.joss.theoj.org/reviewers/3741) and [James Chapman](https://reviewers.joss.theoj.org/reviewers/3035). Thanks!!
