### 2025-03-27 #7954 @editorialbot
[REVIEW]: AniSOAP: Machine Learning Representations for Coarse-grained and Non-spherical Systems
**Submitting author:** @rosecers (<a href="http://orcid.org/0000-0003-4515-3441">Rose K. Cersonsky</a>)
**Repository:** https://github.com/cersonsky-lab/AniSOAP/
**Branch with paper.md** (empty if default branch):
**Version:** v1
