### 2025-04-04 cersonsky-lab/AniSOAP#59 @DaniBodor
[joss review]: add documentation
The package currently has very minimal documentation, largely limited to installation instructions.

Please check the [checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation) and [criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation) for adequate levels of documentation for JOSS papers and expand the README to at least cover these minimal requirements.

Creating comprehensive API documentation, e.g. using [mkdocs](https://www.mkdocs.org/) or [sphinx](https://www.sphinx-doc.org/en/master/index.html) would help users adopt your package even better, but does not appear to be part of the requirements of JOSS.
