### 2025-04-04 cersonsky-lab/AniSOAP#59 @DaniBodor
I have copied the checklist here for your convenience, may be useful for you to keep track of updating this:

    - see https://github.com/cersonsky-lab/AniSOAP/issues/58
