### 2025-04-04 #7954 @DaniBodor
> ### Documentation

Documentation is largely lacking at present. I have opened https://github.com/cersonsky-lab/AniSOAP/issues/59 to address this
