### 2025-04-04 #7954 @DaniBodor
> ### Functionality

I will address this when the documentation has been updated, as that will allow me to better judge what exactly the functional claims are.
