### 2025-04-04 cersonsky-lab/AniSOAP#59 @SamTov
I was also going to make an issue for the documentation but I will just add it here. When building using sphinx I was missing:

- sphinx_rtd_theme
- sphinx_gallery
- sphinx_tabs

These can be added to a dev_requirements.txt or as part of a pyroject.toml. In saying that, setting up a GitHub pages or RTD site would be helpful for people to see your docs without having to build them from scratch.
