### 2025-04-07 4cde1db
Add link to readthedocs to the Readme.
