+++ README.md
@@ -12,6 +12,9 @@ AniSOAP
 While technically complete, AniSOAP is in beta mode, and subject to new changes regularly. 
 Please use with caution as we iron out some of the finer details.
 
+## Documentation
+
+Please read our latest documentation, containing examples and API usage here: [https://anisoap.readthedocs.io/en/latest/](https://anisoap.readthedocs.io/en/latest/)
 ## Dependencies
 
 Before installing anisoap, please make sure you have at least the
