### 2025-04-08 e12a2d4
moved prototype notebooks to its own folder
