+++ notebooks/prototypes/README.md
@@ -1,0 +1,2 @@
+# Prototype Notebooks
+These are old archival notebooks that contain early iterations of the anisoap functions before they were implemented in structured python code.
