### 2025-04-08 #7954 @arthur-lin1027
@SamTov, thank you so much for your comments. I'll be addressing your code comments first within their respective issues, and then I'll address improvements to the paper.

@DaniBodor, thank you as well for your comments. I realize that the link to the hosted documentation needs to be included and some things will need to be made more clear within the documentation itself (e.g. community guidelines, updated installation procedures, etc.) all in this issue: https://github.com/cersonsky-lab/AniSOAP/issues/59
