### 2025-04-08 211ad1a
merge all requirements from docs/ and tests/ into single requirements.txt
