+++ requirements.txt
@@ -1,5 +1,17 @@
+--extra-index-url https://luthaf.fr/nightly-wheels/
 featomic
-metatensor
+ase
+coverage[toml]
+pytest
 scipy
-skmatter
+setuptools-rust
 tqdm
+metatensor
+wigners
+matplotlib
+Sphinx
+skmatter
+Sphinx==7.2.6
+sphinx-rtd-theme==2.0.0
+sphinx-gallery  # convert python files into nice documentation
+sphinx-tabs     # tabs for code examples (one tab per language)
