### 2025-04-10 517d7b5
removed duplicate Sphinx req
