+++ requirements.txt
@@ -9,7 +9,6 @@ tqdm
 metatensor
 wigners
 matplotlib
-Sphinx
 skmatter
 Sphinx==7.2.6
 sphinx-rtd-theme==2.0.0
