### 2025-04-10 407ef18
removed duplicate requirements.txt in tests and docs
