+++ docs/requirements.txt
@@ -1,15 +1,0 @@
-# sphinx dependencies
-Sphinx==7.2.6
-sphinx-rtd-theme==2.0.0
-sphinx-gallery  # convert python files into nice documentation
-sphinx-tabs     # tabs for code examples (one tab per language)
-
-
-metatensor
-
-tqdm
-chemfiles
-matplotlib
-skmatter
-ase
-featomic @ git+https://github.com/Luthaf/featomic
