### 2025-04-10 088ea78
Updated installation steps
