### 2025-04-10 90ef869
modified Makefile and included docs build in installation.rst
