### 2025-04-11 9ac7a2c
Improved installation.rst
