### 2025-04-11 d847181
Small update installation.rst
