+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "anisoap"
 version = "0.0.0"
-requires-python = ">=3.8"
+requires-python = "3.9, 3.10"
 authors = [
   {name = "Arthur Lin", email = "alin62@wisc.edu"},
   {name = "Kevin Kazuki Huguenin-Dumittan"},
@@ -22,7 +22,6 @@ classifiers = [
   "Natural Language :: English",
 
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",  
 ]
