### 2025-04-13 089990a
updated pyproject.toml with exact packages
