### 2025-04-13 73ec11c
removed version number from scipy
