### 2025-04-13 ab757d2
removed version number from matplotlib
