+++ .github/workflows/tests.yml
@@ -22,7 +22,6 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
           pip install .
       - name: Run tests
         run: |
