### 2025-04-13 aa3a2c6
Update tests.yml -- added 3.11, 3.12, 3.13 into valid versions.
