### 2025-04-13 28f4157
updated pyproject.toml with new classifiers
