+++ pyproject.toml
@@ -23,7 +23,9 @@ classifiers = [
 
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",  
+  "Programming Language :: Python :: 3.10", 
+  "Programming Language :: Python :: 3.11", 
+  "Programming Language :: Python :: 3.12",  
 ]
 dependencies = [
   "featomic==0.6.1",
