### 2025-04-14 a754ff4
improved installation.rst
