### 2025-04-14 c53e7d0
removed dependency specifiers in pyproject.toml
