### 2025-04-14 73b5e8b
improved installation.rst
