### 2025-04-14 53a6d8d
slight change to installation.rst
