### 2025-04-14 e7937ee
slight change to installation.rst
