### 2025-04-26 5c07661
Removed c_q restriction from ellipsoidal_density_projection.py
