### 2025-04-28 ca381c4
First draft of CONTRIBUTING.md
