+++ CONTRIBUTING.md
@@ -1,0 +1,44 @@
+# Contributing to AniSOAP
+Thank you for your interest in contributing to AniSOAP! We welcome contributions from the community to help make this package better for molecular simulations and coarse-graining applications.
+
+## How to Contribute
+There are several ways you can contribute to the AniSOAP project:
+
+1. ### Report Bugs or Request Features
+  * If you find a bug or would like to request a new feature, please open an issue in the GitHub Issues tracker.
+  * When reporting a bug, please include details such as:
+  * Steps to reproduce the issue
+  * Any error messages or logs
+  * Version of AniSOAP you are using
+
+2. ### Contribute Code
+  * Fork the repository and create a new branch for your changes.
+  * Make your changes, ensuring they align with the style and conventions used in the project. We use `black` to lint the source code and `isort` to ensure imports are clean. 
+  * Run tests to make sure your code doesn’t break existing functionality.
+  * Submit a pull request (PR) with a clear description of what you’ve done.
+  * If your changes include new features, please add tests to ensure they work correctly.
+
+3. ### Documentation Improvements
+  * Good documentation is key to the usability of the software. If you notice areas that could be improved or better explained, feel free to submit a PR with updates to the documentation. These can be done using the
+
+4. ### Help with Issues
+  * If you see an issue that you know how to fix, consider submitting a PR or commenting with suggestions.
+
+## Coding Guidelines
+* Write clean, maintainable code that follows Python best practices.
+* Use consistent indentation (4 spaces per level).
+* Add docstrings to functions and classes. We use the Numpy docstring style (https://numpydoc.readthedocs.io/en/latest/format.html).
+
+## Testing
+We use pytest for testing. Before submitting your PR, run the tests to make sure everything works:
+
+``pytest``
+
+## Licensing
+By contributing to AniSOAP, you agree that your contributions will be licensed under the same open-source license as the project. Please see the `LICENSE` file for more information.
+
+## Questions?
+If you have any questions, feel free to reach out by opening an issue or contacting the maintainers directly.
+
+Thanks for contributing!
+
