### 2025-04-29 be578db
added CITATION.cff
