+++ requirements.txt
@@ -1,4 +1,3 @@
---extra-index-url https://luthaf.fr/nightly-wheels/
 featomic==0.6.1
 ase==3.25.0
 coverage[toml]
