### 2025-05-16 f1a98be
moved tests to where they need to go
