### 2025-05-16 ad949b1
moved tests to where they need to go
