+++ tests/unit-tests/test_moment_generator.py
@@ -1,0 +1,215 @@
+import numpy as np
+import pytest
+from numpy.testing import assert_allclose
+from scipy.special import (
+    comb,
+    gamma,
+)
+
+# Import the different versions of the moment generators
+from anisoap.utils import (
+    assert_close,
+    compute_moments_diagonal_inefficient_implementation,
+    compute_moments_inefficient_implementation,
+    compute_moments_single_variable,
+)
+
+
+class TestMomentsUnivariateGaussian:
+    """
+    Class for testing the correctness of the single variable moments
+    <x^n> = integral x^n exp(-0.5*A*x^2)
+    against analytical expressions or alternative evaluation schemes.
+    """
+
+    maxdegs = [2, 5, 10]
+    sigmas = np.geomspace(0.2, 5, 7)
+    centers = np.linspace(-10, 10, 7)
+
+    # Test the single variable moments against the analytical expression
+    # for the special case of Gaussians centered around zero.
+    @pytest.mark.parametrize("maxdeg", maxdegs)
+    @pytest.mark.parametrize("sigma", sigmas)
+    def test_moments_single_variable_centered(self, sigma, maxdeg):
+        A = 1 / sigma**2
+        moments = compute_moments_single_variable(A, 0, maxdeg)
+
+        moments_exact = np.zeros((maxdeg + 1,))
+        for deg in range(maxdeg + 1):
+            exact_value = 0
+            if deg % 2 == 0:
+                neff = (deg + 1) / 2
+                exact_value = (2 * sigma**2) ** neff * gamma(neff)
+            moments_exact[deg] = exact_value
+
+        assert_allclose(moments, moments_exact, rtol=1e-15, atol=1e-15)
+
+    # Test the single variable moments against the analytical expression
+    # for the general case of a Gaussian centered around an arbitrary
+    # point a.
+    # The reference values are obtained from an alternative approach
+    # to compute the moments using the binomial theorem.
+    @pytest.mark.parametrize("maxdeg", maxdegs)
+    @pytest.mark.parametrize("center", centers)
+    @pytest.mark.parametrize("sigma", sigmas)
+    def test_moments_single_variable_non_centered(self, sigma, center, maxdeg):
+        A = 1 / sigma**2
+
+        # Compute the exact moments for the centered moments
+        centered_moments = np.zeros((maxdeg + 1,))
+        for deg in range(maxdeg + 1):
+            exact_value = 0
+            if deg % 2 == 0:
+                neff = (deg + 1) / 2
+                exact_value = (2 / A) ** neff * gamma(neff)
+            centered_moments[deg] = exact_value
+
+        # Compute the moments from the binomial theorem
+        moments = np.zeros((maxdeg + 1,))
+        for deg in range(maxdeg + 1):
+            # The first term in the binomial expansion
+            # is always the centered moment of the same degree
+            moments[deg] += centered_moments[deg]
+
+            # Get the correction from the centered moment
+            for k in range(deg):
+                moments[deg] -= comb(deg, k) * (-center) ** (deg - k) * moments[k]
+
+        # Compare with the moments obtained from the iterative algorithm
+        # used in the main code
+        moments_from_code = compute_moments_single_variable(A, center, maxdeg)
[…]
