### 2025-05-16 f9a96cf
moved tests to where they need to go
