+++ tests/test_spherical_to_cartesian.py
@@ -1,145 +1,0 @@
-from math import factorial
-
-import numpy as np
-import pytest
-from numpy.testing import assert_allclose
-
-from anisoap.utils import (
-    TrivariateMonomialIndices,
-    assert_close,
-    spherical_to_cartesian,
-)
-
-
-# Generate the exact values of the spherical to
-# Cartesian coordinate transformations for l=0,1,2.
-# For l=0 and l=1, arbitrary degrees in n are supported,
-# while for l=2, only n=0 is supported.
-def spherical_to_cartesian_exact(lmax, num_ns):
-    T = []
-    assert lmax <= 3, "Exact analytical expressions only implemented for lmax<=3"
-    assert (
-        len(num_ns) == lmax + 1
-    ), "Provided list of nmax values has to match number of l values"
-
-    # Initialize the arrays in the same way as the main code
-    for l, nmax in enumerate(num_ns):
-        # Initialize array in which to store all
-        # # coefficients for each l
-        maxdeg = l + 2 * (nmax - 1)
-
-        # Usage T_l[m,n,n0,n1,n2]
-        shape = (2 * l + 1, nmax, maxdeg + 1, maxdeg + 1, maxdeg + 1)
-        T_l = np.zeros((2 * l + 1, nmax, maxdeg + 1, maxdeg + 1, maxdeg + 1))
-        T.append(T_l)
-
-    # Fill in the exact coefficients for l=0
-    for n in range(num_ns[0]):
-        monomial_indices = iter(TrivariateMonomialIndices(deg=n))
-        fac_n = factorial(n)
-        for idx, n0, n1, n2 in monomial_indices:
-            coeff = fac_n / factorial(n0) / factorial(n1) / factorial(n2)
-            T[0][0, n, 2 * n0, 2 * n1, 2 * n2] = coeff
-    if lmax == 0:
-        return T
-
-    # Fill in the exact coefficients for l=1
-    for n in range(num_ns[1]):
-        monomial_indices = iter(TrivariateMonomialIndices(deg=n))
-        fac_n = factorial(n)
-        for idx, n0, n1, n2 in monomial_indices:
-            coeff = fac_n / factorial(n0) / factorial(n1) / factorial(n2)
-            T[1][
-                0, n, 2 * n0, 2 * n1 + 1, 2 * n2
-            ] = coeff  # m=-1 with solid harmonic = y
-            T[1][
-                1, n, 2 * n0, 2 * n1, 2 * n2 + 1
-            ] = coeff  # m=0 with solid harmonic = z
-            T[1][
-                2, n, 2 * n0 + 1, 2 * n1, 2 * n2
-            ] = coeff  # m=1 with solid harmonic = x
-    if lmax == 1:
-        return T
-
-    # Fill in the exact coefficients for l=2 and n=0.
-    assert num_ns[2] == 1, "For l=2, only n=0 is implemented"
-    T[2][0, 0, 1, 1, 0] = np.sqrt(3)  # m=-2
-    T[2][1, 0, 0, 1, 1] = np.sqrt(3)  # m=-1
-    T[2][3, 0, 1, 0, 1] = np.sqrt(3)  # m=+1
-
-    # Coeffs for m=0
-    T[2][2, 0, 2, 0, 0] = -0.5
-    T[2][2, 0, 0, 2, 0] = -0.5
-    T[2][2, 0, 0, 0, 2] = 1
-
-    # Coeffs for m=+2
-    T[2][4, 0, 2, 0, 0] = np.sqrt(3) / 2
-    T[2][4, 0, 0, 2, 0] = -np.sqrt(3) / 2
-
-    return T
-
-
-@pytest.mark.parametrize("lmax", np.arange(5))
[…]
