+++ tests/test_radial_basis.py
@@ -1,238 +1,0 @@
-import numpy as np
-import pytest
-from numpy.testing import assert_allclose
-from scipy.spatial.transform import Rotation
-
-# internal imports
-from anisoap.representations import MonomialBasis, GTORadialBasis, radial_basis
-
-
-class TestNumberOfRadialFunctions:
-    """
-    Test that the number of radial basis functions is correct.
-    """
-
-    def test_radial_functions_n5(self):
-        basis_gto = MonomialBasis(max_angular=5, cutoff_radius=5.0)
-        num_ns = basis_gto.get_num_radial_functions()
-
-        # Compare against exact results
-        num_ns_exact = [3, 3, 2, 2, 1, 1]
-        assert len(num_ns) == len(num_ns_exact)
-        for l, num in enumerate(num_ns):
-            assert num == num_ns_exact[l]
-
-    def test_radial_functions_n6(self):
-        basis_gto = MonomialBasis(max_angular=6, cutoff_radius=5.0)
-        num_ns = basis_gto.get_num_radial_functions()
-
-        # Compare against exact results
-        num_ns_exact = [4, 3, 3, 2, 2, 1, 1]
-        assert len(num_ns) == len(num_ns_exact)
-        for l, num in enumerate(num_ns):
-            assert num == num_ns_exact[l]
-
-    def test_radial_functions_n7(self):
-        basis_gto = MonomialBasis(max_angular=6, max_radial=5, cutoff_radius=5.0)
-        num_ns = basis_gto.get_num_radial_functions()
-
-        # We specify max_radial so it's decoupled from max_angular.
-        max_ns_exact = [5, 5, 5, 5, 5, 5, 5]
-        assert len(num_ns) == len(max_ns_exact)
-        for l, num in enumerate(num_ns):
-            assert num == max_ns_exact[l] + 1
-
-    def test_radial_functions_n8(self):
-        basis_gto = MonomialBasis(
-            max_angular=6,
-            max_radial=[1, 2, 3, 4, 5, 6, 7],
-            cutoff_radius=5.0,
-        )
-        num_ns = basis_gto.get_num_radial_functions()
-
-        # We specify max_radial so it's decoupled from max_angular.
-        max_ns_exact = [1, 2, 3, 4, 5, 6, 7]
-        assert len(num_ns) == len(max_ns_exact)
-        for l, num in enumerate(num_ns):
-            assert num == max_ns_exact[l] + 1
-
-
-class TestBadInputs:
-    """
-    Class for testing if radial_basis fails with bad inputs
-    """
-
-    DEFAULT_HYPERS = {
-        "max_angular": 10,
-        "radial_gaussian_width": 5.0,
-        "cutoff_radius": 1.0,
-    }
-    test_hypers = [
-        # [
-        #     {**DEFAULT_HYPERS, "radial_gaussian_width": 5.0, "max_radial": 3},
-        #     ValueError,
-        #     "Only one of max_radial or radial_gaussian_width can be independently specified",
-        # ],
-        [
-            {
-                **DEFAULT_HYPERS,
-                "radial_gaussian_width": 5.0,
-                "max_radial": [1, 2, 3],
-            },  # default max_angular = 10
-            ValueError,
-            "If you specify a list of number of radial components, this list must be of length 11. Received 3.",
-        ],
-        [
-            {**DEFAULT_HYPERS, "radial_gaussian_width": 5.0, "max_radial": "nonsense"},
-            ValueError,
-            "`max_radial` must be None, int, or list of int",
-        ],
-        [
-            {
-                **DEFAULT_HYPERS,
-                "radial_gaussian_width": 5.0,
[…]
