### 2025-05-16 89e9c05
Resolving Documentation-related Issues (#63)
* Add link to readthedocs to the Readme.

* moved prototype notebooks to its own folder

* merge all requirements from docs/ and tests/ into single requirements.txt

* removed duplicate Sphinx req

* removed duplicate requirements.txt in tests and docs

* Updated installation steps

* create refined requirements.txt

* Update tests.yml

* modified Makefile and included docs build in installation.rst

* modified .readthedocs.yaml to add other requirments

* Improved installation.rst

* Small update installation.rst

* updated pyproject.toml

* update
[…]
