+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,34 @@
+# AniSOAP Community Guidelines
+
+At AniSOAP, we strive to create an inclusive, collaborative, and respectful environment for all contributors. These guidelines set expectations for how we interact, collaborate, and work together to improve the project. Whether you’re submitting a pull request, discussing an issue, or simply using the package, we expect everyone to engage in a manner that fosters a positive experience for all participants.
+
+## Be Respectful
+Treat all members of the community with respect. Personal attacks, harassment, or any form of disrespectful communication will not be tolerated. We value diverse perspectives and encourage thoughtful, constructive dialogue.
+
+## Be Inclusive
+We welcome contributors from all backgrounds, skill levels, and experience. Our goal is to build a welcoming and supportive community. If you're new to open source or unsure how to get started, don’t hesitate to ask for help! We encourage questions and feedback from everyone.
+
+## Assume Good Intent
+Disagreements may arise, but we should always approach them with the assumption that everyone is acting in good faith. If you disagree with someone’s approach or perspective, respond constructively and suggest alternatives instead of dismissing others' ideas.
+
+<!-- Stay On-Topic
+Keep discussions focused on the topic at hand. Whether it’s an issue, a pull request, or a feature request, stay relevant to the subject and avoid unnecessary off-topic discussions. This ensures that conversations remain productive and useful for everyone.
+
+No Spam or Self-Promotion
+Please refrain from using the project for personal promotion, spamming, or irrelevant content. Contributions should add value to the project and not be used as a platform for external advertising.
+
+Use Clear Communication
+Clear communication is key to a healthy community. When reporting issues or discussing features, try to be as detailed as possible. Be kind, patient, and open-minded in your interactions. If something isn't clear, ask for clarification.
+
+Report Issues, Not People
+If you encounter inappropriate behavior, please report it directly to the maintainers by opening an issue or contacting the project leads privately. We take all reports seriously and will address them as needed to maintain a safe and respectful environment.
+
+Collaborate Openly
+We encourage you to share your ideas, feedback, and solutions with the community. Collaboration is at the heart of open-source projects, and working together makes AniSOAP better for everyone. Share your knowledge, ask for advice, and help improve the project where you can.
+ -->
+<!-- Code of Conduct
+We follow a Code of Conduct (which can be found here) to help ensure a positive and respectful atmosphere in all community spaces related to AniSOAP. -->
+
+<!-- Help Make the Community Better
+If you notice any areas where we could improve the community experience or have ideas to enhance collaboration, we welcome your input. The more feedback we receive, the better we can make AniSOAP a thriving open-source project. -->
+
