+++ requirements.txt
@@ -1,5 +1,11 @@
-featomic
-metatensor
+featomic==0.6.1
+ase==3.25.0
+coverage[toml]
+pytest==8.3.5
 scipy
-skmatter
-tqdm
+setuptools-rust==1.11.1
+tqdm==4.67.1
+metatensor==0.2.0
+wigners==0.3.0
+matplotlib
+skmatter==0.2.0

[…]
