### 2025-05-28 cersonsky-lab/AniSOAP#70 @DaniBodor
[joss_review]: fix reference
when test running the notebook, these three references (in cell 3 of Invariances_of_powerspectrum.ipynb) to `ellipsoids.xyz` do not point to the correct folder location, and should be adjusted to `read("../ellipsoids.xyz", "0:2")`

https://github.com/cersonsky-lab/AniSOAP/blob/89e9c0557b31414adf7059b34ac5ebdd43418841/notebooks/invariances_example/Invariances_of_powerspectrum.ipynb#L45-L47
