### 2025-05-28 cersonsky-lab/AniSOAP#70 @DaniBodor
Similarly, pytest currently crashes looking for the same files

> __________________________ ERROR collecting notebooks/example01_invariances_of_powerspectrum_test.py __________________________
> notebooks/example01_invariances_of_powerspectrum_test.py:31: in <module>
>     frames = read("ellipsoids.xyz", "0:2")
> ../../miniconda3/envs/testANISOAP/lib/python3.11/site-packages/ase/io/formats.py:808: in read
>     format = format or filetype(filename, read=isinstance(filename, str))
> ../../miniconda3/envs/testANISOAP/lib/python3.11/site-packages/ase/io/formats.py:983: in filetype
>     fd = open_with_compression(filename, 'rb')
> ../../miniconda3/envs/testANISOAP/lib/python3.11/site-packages/ase/io/formats.py:612: in open_with_compression
>     return open(filename, mode)
> E   FileNotFoundError: [Errno 2] No such file or directory: 'ellipsoids.xyz'
> =================================================== short test summary info ===================================================
> ERROR notebooks/example01_invariances_of_powerspectrum_test.py - FileNotFoundError: [Errno 2] No such file or directory: 'ellipsoids.xyz'
