### 2025-05-28 #7954 @DaniBodor
Apologies for the delay from my side.

@arthur-lin1027 , thank you for including the link to the documentation. Many things are nicely covered there. However, I've opened up a few new (minor) issues I think will be very helpful for users to get started with your code.

As to the functionality of the code, as mentioned when I accepted to review the code, I do not know the domain well enough to evaluate the point "Have the functional claims of the software been confirmed". The best I can do is test run the code snippets and examples given and point out whether anything is broken or clearly wrong (e.g. I've added an issue regarding incorrect file paths in some examples).
