### 2025-06-05 6cc8c01
Fixing references in notebook
