+++ notebooks/example01_invariances_of_powerspectrum_test.py
@@ -8,7 +8,6 @@ This example demonstrates:
 2. How to convert ellipsoidal frames to AniSOAP vectors.
 3. How to create ellipsoidal frames with ``ase.Atoms``.
 """
-
 from ase.io import read
 from ase import Atoms
 
@@ -28,9 +27,9 @@ import matplotlib.pyplot as plt
 # %%
 # Read the first two frames of ellipsoids.xyz, which represent coarse-grained benzene molecules.
 
-frames = read("./ellipsoids.xyz", "0:2")
-frames_translation = read("./ellipsoids.xyz", "0:2")
-frames_rotation = read("./ellipsoids.xyz", "0:2")
+frames = read("notebooks/ellipsoids.xyz", "0:2")
+frames_translation = read("notebooks/ellipsoids.xyz", "0:2")
+frames_rotation = read("notebooks/ellipsoids.xyz", "0:2")
 
 print(f"{len(frames)=}")  # a list of atoms objects
 print(f"{frames[0].arrays=}")
