+++ notebooks/example02_learn_benzene.py
@@ -30,8 +30,8 @@ import pickle
 lmax = 9
 nmax = 6
 
-atom_frames = read("benzenes.xyz", ":")  # all atom frames, containing benzene energies
-frames = read("ellipsoids.xyz", ":")  # ellipsoid frames
+atom_frames = read("notebooks/benzenes.xyz", ":")  # all atom frames, containing benzene energies
+frames = read("notebooks/ellipsoids.xyz", ":")  # ellipsoid frames
 energies = np.array([aframe.info["energy_pa"] for aframe in atom_frames])
 energies = np.reshape(
     energies, (-1, 1)
