+++ notebooks/example02_learn_benzene.py
@@ -30,7 +30,9 @@ import pickle
 lmax = 9
 nmax = 6
 
-atom_frames = read("notebooks/benzenes.xyz", ":")  # all atom frames, containing benzene energies
+atom_frames = read(
+    "notebooks/benzenes.xyz", ":"
+)  # all atom frames, containing benzene energies
 frames = read("notebooks/ellipsoids.xyz", ":")  # ellipsoid frames
 energies = np.array([aframe.info["energy_pa"] for aframe in atom_frames])
 energies = np.reshape(
