### 2025-06-05 015efd0
Added link to contributing guidelines in README.
