### 2025-06-05 d911053
Removed dependency list and added mention of Rust to README.
