+++ docs/source/installation.rst
@@ -10,7 +10,6 @@ A portion of AniSOAP is written in Rust, so installing the Rust compiler (rustc)
 The rest of the python dependencies are listed below and are all available on PyPI. **You do not have to install these dependencies individually**; installing the AniSOAP package will automatically install the dependencies. We recommend using an environment manager like conda (https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html) to create a clean working environment for AniSOAP.
 
 * `Python: 3.9, 3.10, 3.11, 3.12 supported <https://www.python.org/downloads/>`_
-* `numPy <https://numpy.org/install/>`_
 * `sciPy <https://scipy.org/install/>`_
 * `Atomic Simulation Environment (ASE) <https://wiki.fysik.dtu.dk/ase/install.html>`_
 * `Metatensor <https://docs.metatensor.org/latest/index.html>`_
