+++ anisoap/asecg/CGRep.py
@@ -1,0 +1,337 @@
+#%%
+import numpy as np
+from ase import Atom, Atoms
+from ase.io import read, write
+from ase.io.trajectory import Trajectory
+from featomic import SoapPowerSpectrum
+import metatensor 
+import copy
+from skmatter.preprocessing import StandardFlexibleScaler
+import matplotlib.pyplot as plt
+
+#%%
+# Create Dataclass for CG'ed atom
+from dataclasses import dataclass
+
+@dataclass
+class CGInfo():
+    """
+    Stores information about the indices to convert to CG Beads.
+    """
+    cg_indices: list[int] | list[bool]
+    name: str
+    symbol: str
+
+@dataclass
+class CGFrame():
+    frame: Atoms 
+    cg_indices: list[list]
+
+@dataclass
+class CGSystem():
+    cg_beads: list[CGFrame]
+    rep: metatensor.TensorMap
+    gfre: float
+
+# Helper Functions
+def get_frames(cg_beads:list[CGFrame]) -> list[Atoms]:
+    return [cg_bead.frame for cg_bead in cg_beads]
+
+
+# To Coarsen:
+# def coarsen(AA frame, indices to CG):
+    # Return CG'ed bead.
+
+# def find_rigid(CG frame):
+    # Returns bead indices within CG frames to CG
+
+# def get_AA_indices(bead_indices):
+    # Loop through each bead_index
+    # Find the coresponding AA indices. 
+    # 
+
+import numpy as np
+import ase 
+from scipy.spatial.transform import Rotation as R 
+from skmatter.metrics import global_reconstruction_error as GRE
+from itertools import combinations_with_replacement
+
+def get_center_of_mass(frame: Atoms):
+    # For unwrapped systems, set mic=False.
+    # For wrapped systems, USUALLY, mic=True, but not always.
+    displacements = frame.get_all_distances(mic=False, vector=True)  # vectors is a nxnx3 array, it's the displacement from each point to all other points
+    positions = frame[0].position + displacements[0]
+    masses = frame.get_masses()
+    com = masses @ positions / masses.sum()
+    return com 
+
+def get_moments_of_inertia(frame: Atoms, vectors=False):
+    """
+    COPIED FROM ASE Source: https://wiki.fysik.dtu.dk/ase/_modules/ase/atoms.html#Atoms.get_moments_of_inertia
+    MODIFIED TO Include MINIMUM IMAGE CONVENTION. 
+    Get the moments of inertia along the principal axes.
+
+    The three principal moments of inertia are computed from the
+    eigenvalues of the symmetric inertial tensor. Periodic boundary
+    conditions are ignored. Units of the moments of inertia are
+    amu*angstrom**2.
+    """
+    # For unwrapped systems, set mic=False.
+    # For wrapped systems, USUALLY, mic=True, but not always.
+    displacements = frame.get_all_distances(mic=False, vector=True)  # vectors is a nxnx3 array, it's the displacement from each point to all other points
+    # com = frame.get_center_of_mass()   # ASE's incorrect center of mass without mic. 
+    com = get_center_of_mass(frame)
+    positions = frame[0].position + displacements[0] 
+    positions -= com  # translate center of mass to origin
+    masses = frame.get_masses()
+
+    # Initialize elements of the inertial tensor
+    I11 = I22 = I33 = I12 = I13 = I23 = 0.0
+    for i in range(len(frame)):
+        x, y, z = positions[i]
+        m = masses[i]
+
[…]
