### 2025-06-05 d97b5ff
Added introduction to README. (#73)
