### 2025-06-06 13b2119
Update installation.rst, removed the list of dependencies entirely and referred them to the cargo.toml file.
