### 2025-06-06 7e7fadd
Update installation.rst
