### 2025-06-06 014c493
66 joss review clarify installation instructions (#74)
* Removed dependency list and added mention of Rust to README.

* Removed mention of dependencies not listed in requirements.txt.

* Update installation.rst, removed the list of dependencies entirely and referred them to the cargo.toml file.

* Update installation.rst

---------

Co-authored-by: Lucas Ortengren <ortengren@wisc.edu>
