### 2025-06-06 46d3f72
Added links to the Gallery Header
