### 2025-06-06 848fd77
Added links to the Gallery Header
