### 2025-06-06 1dedb3c
reverted references back to being relative to the notebooks folder
