### 2025-06-07 0a8a0dc
slightly modified the powerspectrum, and deleted Seonwoo's original one as it's largely incorporated in mine
