### 2025-06-07 cersonsky-lab/AniSOAP#70 @arthur-lin1027
I'm going to reopen this issue and link it to my PR here.

> when test running the notebook, these three references (in cell 3 of Invariances_of_powerspectrum.ipynb) to `ellipsoids.xyz` do not point to the correct folder location, and should be adjusted to `read("../ellipsoids.xyz", "0:2")`
>
> [AniSOAP/notebooks/invariances_example/Invariances_of_powerspectrum.ipynb](https://github.com/cersonsky-lab/AniSOAP/blob/89e9c0557b31414adf7059b34ac5ebdd43418841/notebooks/invariances_example/Invariances_of_powerspectrum.ipynb#L45-L47)
>
> Lines 45 to 47 in [89e9c05](/cersonsky-lab/AniSOAP/commit/89e9c0557b31414adf7059b34ac5ebdd43418841)
>
>  "ell_frames = read(\"ellipsoids.xyz\", \"0:2\")\n",
>  "ell_frames_translation = read(\"ellipsoids.xyz\", \"0:2\")\n",
>  "ell_frames_rotation = read(\"ellipsoids.xyz\", \"0:2\")\n",

I just deleted this notebook as it's deprecated.
