### 2025-06-07 cersonsky-lab/AniSOAP#70 @arthur-lin1027
> and `notebooks/example02_learn_benzene.py` cannot find `benzenes.xyz`

If you run this in the notebooks folder it should -- these scripts aren't tests and hence should not be run using pytest in the root folder.
