### 2025-06-08 961f897
Update requirements.txt
Using older version of scikit-learn to remain compatible with skmatter=0.2.0
