### 2025-06-09 #7954 @arthur-lin1027
Hi @DaniBodor and @mbobra,
Thank you for your suggestions. I've incorporated nearly all of @DaniBodor's feedback (see: https://github.com/cersonsky-lab/AniSOAP/pull/76, https://github.com/cersonsky-lab/AniSOAP/pull/75, https://github.com/cersonsky-lab/AniSOAP/pull/74, https://github.com/cersonsky-lab/AniSOAP/pull/73, https://github.com/cersonsky-lab/AniSOAP/pull/72) Note that regarding the comment

> I've added an issue regarding incorrect file paths in some examples

These scripts were meant to be run from its own folder, not from the root AniSOAP folder. Hence, the original file paths were correct. I've also removed extraneous/outdated files that were a cause of confusion, hopefully things are clear now.

My outstanding items right now are:
1. Figure out software versioning.
2. Address @SamTov's comments about including more of the software related topics into the paper.

With those additional fixes, I'm hoping we can push this through!

Thanks for all your review and patience.

Sincerely,
Arthur
