### 2025-06-10 cersonsky-lab/AniSOAP#70 @DaniBodor
Thanks for these improvements.

I still have 2 minor comments about this, up to you whether you feel it's worthwhile addressing:

> > and `notebooks/example02_learn_benzene.py` cannot find `benzenes.xyz`
>
> If you run this in the notebooks folder it should -- these scripts aren't tests and hence should not be run using pytest in the root folder.

I see what you mean. if I enter the notebook folder and then run the file, it does indeed work. However, if I am one folder higher (which is usually my default working directory) in the structure and run `python notebooks/example01_invariances_of_powerspectrum_test.py`, then it does not work. This can easily be remedied by using a path relative to file rather than relative to the working directory using `os.path.abspath(__file__)` or `pathlib.Path(__file__).resolve().parent`

Second, using `test` as the start or end of a python filename (as for `notebooks/example01_invariances_of_powerspectrum_test.py`) will automatically trigger pytest to execute it if you just run `pytest`. In your documentation, you do specify that one should run `pytest tests/.`. However, some users (me included) may be used to run `pytest` by default and not notice this instruction and be surprised that this leads to what appears to be failed tests or images popping up (if using my suggestion above).
