### 2025-06-10 #7954 @DaniBodor
My concerns are addressed.  Thank you for your patience with my nitpicky comments :) I do think some of these smaller changes will help new users get on their way quicker. I added one more small [comment](https://github.com/cersonsky-lab/AniSOAP/issues/70#issuecomment-2959289919) to one of the issues you closed, but this does not need to stand in the way of acceptance.

> As to the functionality of the code, as mentioned when I accepted to review the code, I do not know the domain well enough to evaluate the point "Have the functional claims of the software been confirmed". The best I can do is test run the code snippets and examples given and point out whether anything is broken or clearly wrong (e.g. I've added an issue regarding incorrect file paths in some examples).

@mbobra do you want me to check this box nonetheless, or do I leave it unchecked? Do I need to do anything to formally mark my review as complete/accepted?
