+++ paper/paper.md
@@ -39,11 +39,11 @@ bibliography: paper.bib
 
 # Summary
 
-`AniSOAP` is a package that creates Machine Learning (ML) representations of non-spherical particle configurations; these representations can then be used in ML-driven simulations and analyses. This generalization of existing spherical ML representations therefore aims to bridge the gap between two scientific communities: The machine-learned atomistic simulation community, whose primary concern is obtaining fast and (quantum) accurate descriptions of the complex interactions occuring between (spherical) atoms, and the coarse-grained and colloid modeling community, whose primary concern is understanding emergent behavior of macroscopic particles with (plausibly) complex geometries. `AniSOAP` provide a common framework to answer scientific questions at the intersection of these two fields.
+`AniSOAP` is a package that translates coarse-grained molecular configurations into tensorial representations that are ideal for supervised machine-learning models of thermodynamic quantities and unsupervised data-driven analyses. This generalization of existing spherical ML representations therefore aims to bridge the gap between two scientific communities: The machine-learned atomistic simulation community, whose primary concern is obtaining fast and (quantum) accurate descriptions of the complex interactions occuring between (spherical) atoms, and the coarse-grained and colloid modeling community, whose primary concern is understanding emergent behavior of macroscopic particles with (plausibly) complex geometries. `AniSOAP` provide a common framework to answer scientific questions at the intersection of these two fields.
 
 # Statement of need
 
-Machine learning (ML) has greatly advanced atomistic molecular dynamics (MD), enabling 1. Quick and quantum-accurate MD simulations, and 2. Robust techniques to analyze simulation results. Key to these advancements are the increasingly sophisticated strategies used to *featurize* atomistic environments to sensitively differentiate different configurations; this has enabled supervised, semisupervised, and unsupervised studies across a wide variety of chemical spaces[@behler_atom-centered_2011; @bartok_representing_2013; @de_comparing_2016; @cersonsky_data-driven_2023] (more citations?). However, these techniques are often limited to atomistic resolution, as coarse-grained entities (``particles'' or groups of atoms) are not spherical and thus cannot be fully resolved with many atom-inspired ML representations. In these contexts, where molecules or particles have complex, anisotropic geometry, it is important to resolve the orientation-dependence of their interactions with neighboring particles. 
+Machine learning (ML) has greatly advanced atomistic molecular dynamics (MD), enabling 1. Quick and quantum-accurate MD simulations, and 2. Robust techniques to analyze simulation results. Key to these advancements are the increasingly sophisticated strategies and software used to *featurize* atomistic environments to sensitively differentiate different configurations; this has enabled supervised, semisupervised, and unsupervised studies across a wide variety of chemical spaces[@behler_atom-centered_2011; @bartok_representing_2013; @de_comparing_2016; @cersonsky_data-driven_2023]. However, these techniques are often limited to atomistic resolution, as coarse-grained entities (``particles'' or groups of atoms) are not spherical and thus cannot be fully resolved with many atom-inspired ML representations. In these contexts, where molecules or particles have complex, anisotropic geometry, it is important to resolve the orientation-dependence of their interactions with neighboring particles. 
 
[…]
