+++ paper/paper.bib
@@ -1,3 +1,14 @@
+@article{dscribe,
+  author = {Himanen, Lauri and J{\"a}ger, Marc O.~J. and Morooka, Eiaki V. and Federici Canova, Filippo and Ranawat, Yashasvi S. and Gao, David Z. and Rinke, Patrick and Foster, Adam S.},
+  title = {{DScribe: Library of descriptors for machine learning in materials science}},
+  journal = {Computer Physics Communications},
+  volume = {247},
+  pages = {106949},
+  year = {2020},
+  doi = {10.1016/j.cpc.2019.106949},
+  url = {https://doi.org/10.1016/j.cpc.2019.106949},
+  issn = {0010-4655}
+}
 
 @article{bartok_representing_2013,
 	title = {On representing chemical environments},
