+++ notebooks/example03_mixedrep_benzene.py
@@ -1,0 +1,339 @@
+# %%
+"""
+Example 3: Machine-learning benzene with mixed All Atom and CG Representation.
+============================================================
+This example builds upon the previous example, and demonstrates how 
+to create and evaluate mixed representations. 
+
+1. How to read ellipsoidal frames from ``.xyz`` file.
+2. How to convert ellipsoidal frames to AniSOAP vectors.
+3. How to use these frames in machine learning models.
+"""
+# sphinx_gallery_thumbnail_number = 3
+
+import metatensor
+import numpy as np
+from anisoap.representations import EllipsoidalDensityProjection
+from anisoap.utils import ClebschGordanReal, cg_combine, standardize_keys
+from ase.io import read
+from matplotlib import pyplot as plt
+from matplotlib import rc
+from featomic import SoapPowerSpectrum
+from sklearn.decomposition import PCA
+from skmatter.metrics import global_reconstruction_error as GRE
+from sklearn.model_selection import train_test_split
+from sklearn.metrics import root_mean_squared_error as RMSE
+from skmatter.preprocessing import StandardFlexibleScaler
+import pickle
+
+# %%
+# Read the frames
+
+# lmax = 9
+# nmax = 6
+lmax = 2
+nmax = 2
+rcut = 7.0
+
+atom_frames = read(
+    "./benzenes.xyz", ":"
+)  # all atom frames, containing benzene energies
+frames = read("./ellipsoids.xyz", ":")  # ellipsoid frames
+energies = np.array([aframe.info["energy_pa"] for aframe in atom_frames])
+energies = np.reshape(
+    energies, (-1, 1)
+)  # Turn energies into column vector, required for sklearn
+plt.hist(energies, bins=100)
+plt.xlabel("Loaded Energies, eV")
+plt.show()
+
+
+# %%
+# Computing the AniSOAP Vectors
+#
+# * The ideal semiaxes for the ellipsoid are (4, 4, 0.5)
+
+a1, a2, a3 = 4.0, 4.0, 0.5
+for frame in frames:
+    frame.arrays["c_diameter[1]"] = a1 * np.ones(len(frame))
+    frame.arrays["c_diameter[2]"] = a2 * np.ones(len(frame))
+    frame.arrays["c_diameter[3]"] = a3 * np.ones(len(frame))
+
+AniSOAP_HYPERS = {
+    "max_angular": lmax,
+    "max_radial": nmax,
+    "radial_basis_name": "gto",
+    "subtract_center_contribution": True,
+    "rotation_type": "quaternion",
+    "rotation_key": "c_q",
+    "cutoff_radius": rcut,
+    "radial_gaussian_width": 1.5,
+    "basis_rcond": 1e-8,
+    "basis_tol": 1e-3,
+}
+
+calculator = EllipsoidalDensityProjection(**AniSOAP_HYPERS)
+
+x_anisoap_raw = calculator.power_spectrum(frames)
+
+#%%
+# Create an All Atom Representation with the same Hypers
+SOAP_HYPERS = {
+    "cutoff": {
+        "radius": rcut,
+        "smoothing": {"type": "ShiftedCosine", "width": 0.5},
+    },
+    "density": {
+        "type": "Gaussian",
+        "width": 0.3,
+    },
+    "basis": {
[…]
