### 2025-06-10 ac63925
naive concatenation -- pass 1
