### 2025-06-10 f9e0d88
Revert "naive concatenation -- pass 1" go back to original commit
This reverts commit ac63925c342a5817d116c08c4b869bf179a1edf6.
