+++ notebooks/example03_mixedrep_benzene.py
@@ -226,18 +226,33 @@ print(f"SOAP_CG Train R^2: {lr.score(x_train_soapcg, y_train):.3f}")
 print(f"SOAP_CG Train RMSE: {RMSE(energies[i_train], energies_train_pred_soapcg)}") 
 print(f"SOAP_CG Test R^2: {lr.score(x_test_soapcg, y_test):.3f}")
 print(f"SOAP_CG Test RMSE: {RMSE(energies[i_test], energies_test_pred_soapcg)}") 
+
+#%%
+def mixed_kernel(X1, X2):
+    # Each are rows of a concatenated matrix
+    x_soap1, x_anisoap1 = X1[:21], X1[21:]
+    x_soap2, x_anisoap2 = X2[:21], X2[21:]
+    return np.dot(x_soap1, x_soap2) + np.dot(x_anisoap1, x_anisoap2)
 # %%
 # Create mixed AniSOAP/SOAP representation
-alpha = 1
-x_train_mix_raw = np.hstack((alpha * x_train_soap, (1-alpha) * x_train_anisoap))
-x_train_mix = StandardFlexibleScaler(column_wise=False).fit_transform(x_train_mix_raw)
+from skmatter.preprocessing import KernelNormalizer
+alpha = 0.25
+Kaa = x_train_soap @ x_train_soap.T
+Kcg = x_train_anisoap @ x_train_anisoap.T
+Kicg = x_train_soapcg @ x_train_soapcg.T
+Kmix = alpha * Kaa + (1 - alpha) * Kcg
+# x_train_mix_raw = np.hstack((alpha * x_train_soap, (1-alpha) * x_train_anisoap))
+# x_train_mix_raw = alpha * x_train_soap.T @ x_train_soap + (1 - alpha) * x_train_anisoap.T @ x_train_anisoap
+# x_train_mix = StandardFlexibleScaler(column_wise=False).fit_transform(x_train_mix_raw)
+# x_train_mix = 
+lr = KernelRidge(alpha=1e-8, kernel="precomputed")
 
 x_test_mix_raw = np.hstack((alpha * x_test_soap, (1-alpha) * x_test_anisoap))
-# 
+# x_test_mix_raw = alpha * x_test_soap.T @ x_test_soap + (1 - alpha ) * x_test_anisoap.T @ x_test_anisoap
 # x_test_mix_raw = x_test_mix_raw[:, x_test_mix_raw.var(axis=0)>1e-12]
 x_test_mix = StandardFlexibleScaler(column_wise=False).fit_transform(x_test_mix_raw)
 
-lr.fit(x_train_mix, y_train)
+lr.fit(Kmix, y_train)
 
 y_train_pred_mix = lr.predict(x_train_mix)
 y_test_pred_mix = lr.predict(x_test_mix)
@@ -250,6 +265,7 @@ print(f"MixedRep Test R^2: {lr.score(x_test_mix, y_test):.3f}")
 print(f"MixedRep Test RMSE: {RMSE(energies[i_test], energies_test_pred_mix)}") 
 # %%
 from tqdm.auto import tqdm
+from sklearn.decomposition import PCA 
 train_rmses = []
 test_rmses = []
 alphas = np.linspace(0, 1, 101)
@@ -260,16 +276,11 @@ for alpha in tqdm(alphas):
     x_test_mix_raw = np.hstack((alpha * x_test_soap, (1-alpha) * x_test_anisoap))
     # x_test_mix_raw = x_test_mix_raw[:, x_test_mix_raw.var(axis=0)>1e-12]
     x_test_mix = StandardFlexibleScaler(column_wise=False).fit_transform(x_test_mix_raw)
- 
+
     lr.fit(x_train_mix, y_train)
[…]
