### 2025-06-10 ba08018
made working kernel and it seems to work, still need to normalize and put in for loop
