### 2025-06-11 2f8a4ba
add kernel normalizer - works for Train but not for test
