### 2025-06-11 3789428
commented out some extraneous hstacking code, KernelNormalizer still not fixed
