### 2025-06-11 d687fef
made kernelnormalizer work, deleetd extra comments
