### 2025-06-11 ea43c97
added icg kernel normalizer
