### 2025-06-11 c33fc69
Alpha Sweep with Superimposed Banana Plots of AniCG and IsoCG
