+++ notebooks/example03_mixedrep_benzene.py
@@ -293,25 +293,59 @@ print(f"MixedRepIsotropic Test R^2: {kr_icg.score(Kmix_icg_test, y_test):.3f}")
 print(f"MixedRepIsotropic Test RMSE: {RMSE(energies[i_test], energies_test_pred_mix_icg)}") 
 # %%
 from tqdm.auto import tqdm
-from sklearn.decomposition import PCA 
 train_rmses = []
 test_rmses = []
+train_rmses_icg = []
+test_rmses_icg = []
 alphas = np.linspace(0, 1, 101)
+Kaa = x_train_soap @ x_train_soap.T
+Kcg = x_train_anisoap @ x_train_anisoap.T
+Kicg = x_train_soapcg @ x_train_soapcg.T
+
+kernel_normalizer_aa = KernelNormalizer()
+kernel_normalizer_cg = KernelNormalizer()
+kernel_normalizer_icg = KernelNormalizer()
+kernel_normalizer_aa.fit(Kaa)
+kernel_normalizer_cg.fit(Kcg)
+kernel_normalizer_icg.fit(Kicg)
+Kaa_norm = kernel_normalizer_aa.transform(Kaa)
+Kcg_norm = kernel_normalizer_cg.transform(Kcg)
+Kicg_norm = kernel_normalizer_icg.transform(Kicg)
+
+Kaa_test = x_test_soap @ x_train_soap.T
+Kcg_test = x_test_anisoap @ x_train_anisoap.T
+Kicg_test = x_test_soapcg @ x_train_soapcg.T 
+
+Kaa_test = kernel_normalizer_aa.transform(Kaa_test)
+Kcg_test = kernel_normalizer_cg.transform(Kcg_test)
+Kicg_test = kernel_normalizer_icg.transform(Kicg_test)
 for alpha in tqdm(alphas):
-    x_train_mix_raw = np.hstack((alpha * x_train_soap, (1-alpha) * x_train_anisoap))
-    x_train_mix = StandardFlexibleScaler(column_wise=False).fit_transform(x_train_mix_raw)
+   
+    Kmix = alpha * Kaa_norm + (1 - alpha) * Kcg_norm
+    Kmix_icg = alpha * Kaa_norm + (1 - alpha) * Kicg_norm
+    Kmix_test = alpha * Kaa_test + (1 - alpha) * Kcg_test
+    Kmix_icg_test = alpha * Kaa_test + (1 - alpha) * Kicg_test 
 
-    x_test_mix_raw = np.hstack((alpha * x_test_soap, (1-alpha) * x_test_anisoap))
-    # x_test_mix_raw = x_test_mix_raw[:, x_test_mix_raw.var(axis=0)>1e-12]
-    x_test_mix = StandardFlexibleScaler(column_wise=False).fit_transform(x_test_mix_raw)
+    kr = KernelRidge(alpha=1e-8, kernel="precomputed")
+    kr_icg = KernelRidge(alpha=1e-8, kernel="precomputed")
 
-    lr.fit(x_train_mix, y_train)
-    y_train_pred_mix = lr.predict(x_train_mix)
-    y_test_pred_mix = lr.predict(x_test_mix)
+    kr.fit(Kmix, y_train)
+    kr_icg.fit(Kmix_icg, y_train)
+
+    y_train_pred_mix = kr.predict(Kmix)
+    y_test_pred_mix = kr.predict(Kmix_test)
     energies_train_pred_mix = y_train_scaler.inverse_transform(y_train_pred_mix.reshape(-1, 1))
     energies_test_pred_mix = y_test_scaler.inverse_transform(y_test_pred_mix.reshape(-1, 1))
+
+    y_train_pred_mix_icg = kr_icg.predict(Kmix_icg)
[…]
