### 2025-06-11 f5df358
replaced featomic for anisoap for spherical representations
