### 2025-06-11 c252b63
use smaller semiradii for isotropic CG rep, results in smaller difference
