### 2025-06-11 e0e7ea2
Add optimization sections to paper.
