### 2025-06-11 a721180
Add section about Github CI
