### 2025-06-12 cfe9d21
highlighted other descriptor software and highlighted our uniqueness
