### 2025-06-12 fe491bb
highlighted other descriptor software and highlighted our uniqueness
