+++ paper/paper.bib
@@ -1,3 +1,30 @@
+@article{librascal_paper,
+  title = {Efficient implementation of atom-density representations},
+  volume = {154},
+  ISSN = {1089-7690},
+  url = {http://dx.doi.org/10.1063/5.0044689},
+  DOI = {10.1063/5.0044689},
+  number = {11},
+  journal = {The Journal of Chemical Physics},
+  publisher = {AIP Publishing},
+  author = {Musil,  Félix and Veit,  Max and Goscinski,  Alexander and Fraux,  Guillaume and Willatt,  Michael J. and Stricker,  Markus and Junge,  Till and Ceriotti,  Michele},
+  year = {2021},
+  month = mar 
+}
+
+@software{librascal,
+  title = {librascal},
+  year = {2021},
+  url = {https://github.com/lab-cosmo/librascal}
+}
+
+@software{quippy,
+  title={QUIP and quippy documentation},
+  author={Bernstein, Noam and Csányi, Gábor and Kermode, James},
+  year={2020},
+  url={https://libatoms.github.io/QUIP/}
+}
+
 @article{dscribe,
   author = {Himanen, Lauri and J{\"a}ger, Marc O.~J. and Morooka, Eiaki V. and Federici Canova, Filippo and Ranawat, Yashasvi S. and Gao, David Z. and Rinke, Patrick and Foster, Adam S.},
   title = {{DScribe: Library of descriptors for machine learning in materials science}},
