+++ paper/paper.bib
@@ -11,7 +11,13 @@
   year = {2021},
   month = mar 
 }
-
+Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan, Davide Tisi, Alexander Goscinski
+@software{featomic,
+  title = {featomic},
+  author = {Fraux, Guillaume and Loche, Philip and Kliavinek, Sergei and Kazuki Huguenin-Dummitan, Kevin and Tisi, Davide and Goscinski, Alexander}
+  year = {2025},
+  url = {https://github.com/metatensor/featomic}
+}
 @software{librascal,
   title = {librascal},
   year = {2021},
