+++ paper/paper.bib
@@ -14,7 +14,7 @@
 Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan, Davide Tisi, Alexander Goscinski
 @software{featomic,
   title = {featomic},
-  author = {Fraux, Guillaume and Loche, Philip and Kliavinek, Sergei and Kazuki Huguenin-Dummitan, Kevin and Tisi, Davide and Goscinski, Alexander}
+  author = {Fraux, Guillaume and Loche, Philip and Kliavinek, Sergei and Kazuki Huguenin-Dummitan, Kevin and Tisi, Davide and Goscinski, Alexander},
   year = {2025},
   url = {https://github.com/metatensor/featomic}
 }
