### 2025-06-12 fb8c27b
added random seed to example03 to make results reproducible
