### 2025-06-15 7404040
incorporate nearly all of JN's suggestions
