+++ paper/paper.bib
@@ -1,3 +1,23 @@
+@article{dusson2022,
+title={Atomic cluster expansion: Completeness, efficiency and stability},
+author={Dusson, Genevieve and Bachmayr, Markus and Cs{'a}nyi, G{'a}bor and Drautz, Ralf and Etter, Simon and van Der Oord, Cas and Ortner, Christoph},
+journal={Journal of Computational Physics},
+volume={454},
+pages={110946},
+year={2022},
+publisher={Elsevier}
+}
+
+@article{nice,
+title={Recursive evaluation and iterative contraction of N-body equivariant features},
+author={Nigam, Jigyasa and Pozdnyakov, Sergey and Ceriotti, Michele},
+journal={The Journal of chemical physics},
+volume={153},
+number={12},
+year={2020},
+publisher={AIP Publishing}
+}
+
 @article{librascal_paper,
   title = {Efficient implementation of atom-density representations},
   volume = {154},
