### 2025-06-15 53056c0
incorporate nearly all of JN's suggestions
