### 2025-06-16 81b641c
incorporated some more changes and deleted 'what is an anisoap vector' section
