### 2025-06-16 0662dc3
incorporated nequip and mace into paper
