+++ paper/paper.bib
@@ -1,3 +1,47 @@
+@article{nequip, 
+  title = {E(3)-equivariant graph neural networks for data-efficient and accurate interatomic potentials},
+  volume = {13},
+  ISSN = {2041-1723},
+  url = {http://dx.doi.org/10.1038/s41467-022-29939-5},
+  DOI = {10.1038/s41467-022-29939-5},
+  number = {1},
+  journal = {Nature Communications},
+  publisher = {Springer Science and Business Media LLC},
+  author = {Batzner,  Simon and Musaelian,  Albert and Sun,  Lixin and Geiger,  Mario and Mailoa,  Jonathan P. and Kornbluth,  Mordechai and Molinari,  Nicola and Smidt,  Tess E. and Kozinsky,  Boris},
+  year = {2022},
+  month = may 
+}
+
+@inproceedings{mace,
+author = {Batatia, Ilyes and Kov\'{a}cs, David P\'{e}ter and Simm, Gregor N. C. and Ortner, Christoph and Cs\'{a}nyi, G\'{a}bor},
+title = {MACE: higher order equivariant message passing neural networks for fast and accurate force fields},
+year = {2022},
+isbn = {9781713871088},
+publisher = {Curran Associates Inc.},
+address = {Red Hook, NY, USA},
+abstract = {Creating fast and accurate force fields is a long-standing challenge in computational chemistry and materials science. Recently, several equivariant message passing neural networks (MPNNs) have been shown to outperform models built using other approaches in terms of accuracy. However, most MPNNs suffer from high computational cost and poor scalability. We propose that these limitations arise because MPNNs only pass two-body messages leading to a direct relationship between the number of layers and the expressivity of the network. In this work, we introduce MACE, a new equivariant MPNN model that uses higher body order messages. In particular, we show that using four-body messages reduces the required number of message passing iterations to just two, resulting in a fast and highly parallelizable model, reaching or exceeding state-of-the-art accuracy on the rMD17, 3BPA, and AcAc benchmark tasks. We also demonstrate that using higher order messages leads to an improved steepness of the learning curves.},
+booktitle = {Proceedings of the 36th International Conference on Neural Information Processing Systems},
+articleno = {830},
+numpages = {14},
+location = {New Orleans, LA, USA},
+series = {NIPS '22}
+}
+
+@article{Cheng2019,
+  title = {Ab initio thermodynamics of liquid and solid water},
+  volume = {116},
+  ISSN = {1091-6490},
+  url = {http://dx.doi.org/10.1073/pnas.1815117116},
+  DOI = {10.1073/pnas.1815117116},
+  number = {4},
+  journal = {Proceedings of the National Academy of Sciences},
+  publisher = {Proceedings of the National Academy of Sciences},
+  author = {Cheng,  Bingqing and Engel,  Edgar A. and Behler,  J\"{o}rg and Dellago,  Christoph and Ceriotti,  Michele},
+  year = {2019},
+  month = jan,
+  pages = {1110–1115}
+}
+
 @article{dusson2022,
 title={Atomic cluster expansion: Completeness, efficiency and stability},
 author={Dusson, Genevieve and Bachmayr, Markus and Cs{'a}nyi, G{'a}bor and Drautz, Ralf and Etter, Simon and van Der Oord, Cas and Ortner, Christoph},
