+++ paper/paper.md
@@ -45,7 +45,7 @@ bibliography: paper.bib
 
 Machine learning (ML) has greatly advanced atomistic molecular dynamics (MD), enabling both quick and quantum-accurate MD simulations and offering powerful tools for analyzing simulation results. Key to these advancements are the increasingly sophisticated strategies and software used to featurize atomistic environments to capture subtle differences between molecular configurations, either explicitly [@behler_atom-centered_2011; @bartok_representing_2013; @drautz_atomic_2019] or implicitly [@mace; @nequip]. These techniques have enabled supervised, semisupervised, and unsupervised studies across a wide variety of chemical spaces[@de_comparing_2016; @Cheng2019; @cersonsky_data-driven_2023]. However, these techniques are largely limited to atomistic resolution, and fall short in reliably describing coarse-grained entities (``particles'' or groups of atoms) that have anisotropic and non-spherical geometries. In such cases, it is essential to resolve the orientation-dependence of their interactions with neighboring particles.
 
-While software exists for creating spherical atomistic descriptors (e.g. DScribe[@dscribe], librascal[@librascal; @librascal_paper], featomic[@featomic]), currently, there are currently no available packages to implement their anisotropic counterparts. In this software, we present the implementation of AniSOAP, an anisotropic generalization of the popular Smooth Overlap of Atomic Positions (SOAP) featurization[@bartok_representing_2013]. SOAP, like other atomistic representations, offers a concise and numerically efficient parameterization of atomistic environments, incorporating correlations of the central atom with up to two of its neighbors. Along with several methods that refine its construction [@nice, @dusson2022], it provides a framework to systematically build higher-order geometric and symmetrized ``fingerprints'' that can be used to model complex interaction potentials and extract machine-learning-enabled insights from data. AniSOAP extends SOAP by additionally incorporating the non-sphericity of particles in terms of particle geometry and orientation. Hence, AniSOAP can be used as a geometrically accurate, high body-order coarse-grained (CG) featurization of molecular and macromolecular systems. As a generalization of SOAP, AniSOAP retains full compatibility with SOAP, and the two representations can be used together to represent molecules at multiple resolutions.
+While software exists for creating spherical atomistic descriptors (e.g. DScribe[@dscribe], librascal[@librascal; @librascal_paper], featomic[@featomic]), currently, there are currently no available packages to implement their anisotropic counterparts. In this software, we present the implementation of AniSOAP, an anisotropic generalization of the popular Smooth Overlap of Atomic Positions (SOAP) featurization[@bartok_representing_2013]. SOAP, like other atomistic representations, offers a concise and numerically efficient parameterization of atomistic environments, incorporating correlations of the central atom with up to two of its neighbors. Along with several methods that refine its construction [@nice; @dusson2022], it provides a framework to systematically build higher-order geometric and symmetrized ``fingerprints'' that can be used to model complex interaction potentials and extract machine-learning-enabled insights from data. AniSOAP extends SOAP by additionally incorporating the non-sphericity of particles in terms of particle geometry and orientation. Hence, AniSOAP can be used as a geometrically accurate, high body-order coarse-grained (CG) featurization of molecular and macromolecular systems. As a generalization of SOAP, AniSOAP retains full compatibility with SOAP, and the two representations can be used together to represent molecules at multiple resolutions.
 
 This technology is motivated by the need to reduce computational complexity in molecular studies, either from a performance-based standpoint (despite high-performance capabilities, many systems cannot be simulated with all-atom resolution in reasonable times) or from a conceptual standpoint, as we may not always want to analyze the behavior of _atoms_, but superatomic entities, such as functional groups. 
[…]
