+++ paper/paper.md
@@ -52,7 +52,7 @@ While CG has made enormous strides in accelerating simulation and analysis, it's
 
 By incorporating anisotropy, AniSOAP extends the advances of atomistic machine learning to coarse-grained and mesoscale systems. With the flexibility of machine learning algorithms and anchored by the concepts underlying the feature-rich SOAP representation, AniSOAP provides a promising conduit for providing data-driven insights to many chemical systems at different length and time-scales.
 
-The main aim of the `AniSOAP` package is to enable the creation of `AniSOAP` feature vectors, which represent systems of ellipsoidal particles. Analogous to how SOAP or ACE create _atom_-centered representations, `AniSOAP` instead creates _particle_-centered representations, where a particle is an anisotropic coarse-grained group of atoms. Specific use-cases of AniSOAP can be seen in our paper here[@lin_expanding_2024].
+The main aim of the `AniSOAP` package is to enable the creation of AniSOAP feature vectors, which represent systems of ellipsoidal particles. Analogous to how SOAP or ACE create _atom_-centered representations, `AniSOAP` instead creates _particle_-centered representations, where a particle is an anisotropic coarse-grained group of atoms. Specific use-cases of AniSOAP can be seen in our paper here[@lin_expanding_2024].
 
 # Implementation details
 The AniSOAP package currently takes in as input a list of frames in the `Atomic Simulation Environment` package[@hjorth_larsen_atomic_2017]. Each frame contains the particles' positions, dimensions, and orientations. If using periodic boundary conditions, the frame also needs to contain the dimensions and orientations of the unit cell. Additional information about each frame can also be stored (e.g. the system energy) and used as a target for supervised ML.
