### 2025-06-19 8d88417
minor grammar fixes
