### 2025-06-21 389241c
Reduced repetitiveness in the paper, thanks to JN
