### 2025-06-21 a7c1584
Beefed up the implementation details section
