+++ paper/paper.md
@@ -55,14 +55,14 @@ The `AniSOAP` package enables the creation of AniSOAP feature vectors, which rep
 # Implementation details
 The AniSOAP package currently takes in as input a list of frames in the `Atomic Simulation Environment` package[@hjorth_larsen_atomic_2017]. Each frame contains the particles' positions, dimensions, and orientations. If using periodic boundary conditions, the frame also needs to contain the dimensions and orientations of the unit cell. Additional information about each frame can also be stored (e.g. the system energy) and used as a target for supervised ML.
 
-With this information, one can construct an `EllipsoidalDensityProjection` object, whose main functionality is to calculate the expansion coefficients of an anisotropic density field in each frame.
-Procedurally, calculating the expansion coefficients amounts to repeatedly and recursively computing high-order moments of an underlying multivariate gaussian, as outlined in[@lin_expanding_2024]. For efficient computation, we have ported these highly-repeated calculations to Rust, a high-performance compiled language.
+With this information, one can construct an `EllipsoidalDensityProjection` object, whose main functionality is to calculate the expansion coefficients of an anisotropic density field in each frame via the `transform` method.
+Procedurally, calculating the expansion coefficients amounts to repeatedly and recursively computing high-order moments of an underlying multivariate gaussian, as outlined in[@lin_expanding_2024]. For efficient computation, we have ported these highly-repeated calculations to Rust, a high-performance compiled language. The intermediate results utilize the `metatensor TensorMap` format[@guillaume_fraux_metatensor_2024], which efficiently stores the AniSOAP featurizations and their associated metadata.
 
-One can take Clebsch-Gordan products of these expansion coefficients to create higher body-order descriptors, and we optimize this step by caching intermediate results with a Least Recently Used (LRU) cache.
+One can take Clebsch-Gordan products of these expansion coefficients to create higher body-order descriptors, and we optimize this step by caching intermediate results with a Least Recently Used (LRU) cache. This functionality is enabled by the `anisoap.metatensor_utils` module in AniSOAP, in particular, the `cg_combine` function.
 
-As many users will be primarily interested in power-spectrum (i.e. 3-body) representations, we provide all the functionality required for these processes, and also provide the convenience method `power_spectrum` to calculate the 3-body descriptors of each frame. 
+As many users will be primarily interested in power-spectrum (i.e. 3-body) representations, we provide all the functionality required for these processes, and also provide the convenience method `power_spectrum` to calculate the 3-body descriptors of each frame. By default, this method returns the featurization as a $n_{\mathrm{samples}}\times n_{\mathrm{features}}$ `numpy` array, which can be used as input into a machine learning algorithm. Alternatively, by setting the keyword argument `mean_over_samples=False`, this method returns a `metatensor TensorMap` object, which contains the power-spectrum representation for each atom in each frame as well as associated metadata. This is a much larger, unaggregated data object that requires more processing before it can be used in an ML algorithm. Examples of the various ways of creating AniSOAP representations can be found in the [following example] (https://anisoap.readthedocs.io/en/latest/auto_examples/example01_invariances_of_powerspectrum_test.html#sphx-glr-auto-examples-example01-invariances-of-powerspectrum-test-py) in our documentation.
 
-The library is thoroughly tested, with unit-tests to test basic functionality, integration-tests to ensure that AniSOAP vectors are calculated correctly, and caching and speed tests to ensure that our aforementioned optimizations yield faster code. These tests are integrated into a Github CI, and we ensure that future features should necessistate additional tests and should pass existing ones.
[…]
