### 2025-06-22 1e0a855
Update GALLERY_HEADER.rst, need to add various ways of creating AniSOAP vectors to example1
