+++ notebooks/GALLERY_HEADER.rst
@@ -6,7 +6,7 @@ Here is a collection of examples that may prove useful. These examples can be vi
 * `Example 1 <https://github.com/cersonsky-lab/AniSOAP/blob/main/notebooks/example01_invariances_of_powerspectrum_test.ipynb>`_
 * `Example 2 <https://github.com/cersonsky-lab/AniSOAP/blob/main/notebooks/example02_learn_benzene.ipynb>`_
 
-The 1st example is a minimal example demonstrating how to create an AniSOAP vector from an ase.Atoms object. It also demonstrates the rotational and translational invariance.
+The 1st example is a minimal example demonstrating the various ways creating an AniSOAP vector from an ase.Atoms object, either via taking manual Clebsch-Gordan products of the expansion coefficients, or via the `power-spectrum` convenience methods. It also demonstrates the rotational and translational invariance.
 
 The 2nd example is an example demonstrating how AniSOAP can be used to coarse-grain benzene and learn benzene crystal energies through linear regression.
 
