+++ profile/README.md
@@ -1,0 +1,4 @@
+# Code and Data used for Profiling AniSOAP Code
+
+Please create data files and processing scripts here that help profile code.
+We may not end up merging everything into the main branch especially if the input data or the outputs are large. 
