+++ paper/paper.md
@@ -60,7 +60,7 @@ Procedurally, calculating the expansion coefficients amounts to repeatedly and r
 
 One can take Clebsch-Gordan products of these expansion coefficients to create higher body-order descriptors, and we optimize this step by caching intermediate results with a Least Recently Used (LRU) cache. This functionality is enabled by the `anisoap.metatensor_utils` module in AniSOAP, in particular, the `cg_combine` function.
 
-As many users will be primarily interested in power-spectrum (i.e. 3-body) representations, we provide all the functionality required for these processes, and also provide the convenience method `power_spectrum` to calculate the 3-body descriptors of each frame. By default, this method returns the featurization as a $n_{\mathrm{samples}}\times n_{\mathrm{features}}$ `numpy` array, which can be used as input into a machine learning algorithm. Alternatively, by setting the keyword argument `mean_over_samples=False`, this method returns a `metatensor TensorMap` object, which contains the power-spectrum representation for each atom in each frame as well as associated metadata. This is a much larger, unaggregated data object that requires more processing before it can be used in an ML algorithm. Examples of the various ways of creating AniSOAP representations can be found in the [following example] (https://anisoap.readthedocs.io/en/latest/auto_examples/example01_invariances_of_powerspectrum_test.html#sphx-glr-auto-examples-example01-invariances-of-powerspectrum-test-py) in our documentation.
+As many users will be primarily interested in power-spectrum (i.e. 3-body) representations, we provide all the functionality required for these processes, and also provide the convenience method `power_spectrum` to calculate the 3-body descriptors of each frame. By default, this method returns the featurization as a $n_{\mathrm{samples}}\times n_{\mathrm{features}}$ `numpy` array, which can be used as input into a machine learning algorithm. Alternatively, by setting the keyword argument `mean_over_samples=False`, this method returns a `metatensor TensorMap` object, which contains the power-spectrum representation for each atom in each frame as well as associated metadata. This is a much larger, unaggregated data object that requires more processing before it can be used in an ML algorithm. Examples of the various ways of creating AniSOAP representations can be found in the examples section of our documentation[@anisoap_documentation].
 
 The library is thoroughly tested and documented, with unit-tests to test basic functionality, integration-tests to ensure that AniSOAP vectors are calculated correctly, and caching and speed tests to ensure that our aforementioned optimizations yield faster code. These tests are integrated into a Github CI, and we ensure that future features should necessistate additional tests and should pass existing ones.
 
