+++ paper/paper.bib
@@ -193,6 +193,16 @@ Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan,
 	file = {Lin et al. - 2024 - Expanding density-correlation machine learning rep.pdf:/Users/alin62/Zotero/storage/HFIZ8QCS/Lin et al. - 2024 - Expanding density-correlation machine learning rep.pdf:application/pdf},
 }
 
+@misc{anisoap_documentation,
+	type = {Documentation},
+	title = {AniSOAP},
+	url = {https://anisoap.readthedocs.io/en/latest/},
+	urldate = {2025-06-23},
+	journal = {Metatensor Documentation},
+	author = {{Arthur Lin} and {Lucas Ortengren} and {Seonwoo Hwang} and {Yong-Cheol Cho} and {Jigyasa Nigam} and {Rose Cersonsky}},
+	year = {2025},
+}
+
 @misc{guillaume_fraux_metatensor_2024,
 	type = {Documentation},
 	title = {Metatensor},
