### 2025-06-23 86d47c2
Modified example01 to include manual descriptor calc
