### 2025-06-23 505185c
updated ipynb and .bib file with proper author
