+++ paper/paper.bib
@@ -199,7 +199,7 @@ Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan,
 	url = {https://anisoap.readthedocs.io/en/latest/},
 	urldate = {2025-06-23},
 	journal = {Metatensor Documentation},
-	author = {{Arthur Lin} and {Lucas Ortengren} and {Seonwoo Hwang} and {Yong-Cheol Cho} and {Jigyasa Nigam} and {Rose Cersonsky}},
+	author = {Lin, Arthur and Ortengren, Lucas and Hwang, Seonwoo and Cho, Yeong-Cheol and Nigam, Jigyasa and Cersonsky, Rose K.},
 	year = {2025},
 }
 
@@ -209,7 +209,7 @@ Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan,
 	url = {https://docs.metatensor.org/latest/index.html},
 	urldate = {2024-09-23},
 	journal = {Metatensor Documentation},
-	author = {{Guillaume Fraux} and {Davide Tisi} and {Philip Loche} and {Joseph W. Abbott} and {Jigyasa Nigam} and {Chiheb Ben Mahmoud}},
+	author = {Fraux, Guillaume and Tisi, Davide and Loche, Philip and Abbott, Joseph W. and Nigam, Jigyasa and Mahmoud, Chiheb Ben},
 	year = {2024},
 }
 
