### 2025-06-23 5e4bd03
updated .bib to remove journal names from Cocumentation.
