+++ paper/paper.bib
@@ -198,7 +198,6 @@ Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan,
 	title = {AniSOAP},
 	url = {https://anisoap.readthedocs.io/en/latest/},
 	urldate = {2025-06-23},
-	journal = {Metatensor Documentation},
 	author = {Lin, Arthur and Ortengren, Lucas and Hwang, Seonwoo and Cho, Yeong-Cheol and Nigam, Jigyasa and Cersonsky, Rose K.},
 	year = {2025},
 }
@@ -208,7 +207,6 @@ Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan,
 	title = {Metatensor},
 	url = {https://docs.metatensor.org/latest/index.html},
 	urldate = {2024-09-23},
-	journal = {Metatensor Documentation},
 	author = {Fraux, Guillaume and Tisi, Davide and Loche, Philip and Abbott, Joseph W. and Nigam, Jigyasa and Mahmoud, Chiheb Ben},
 	year = {2024},
 }
