+++ paper/paper.md
@@ -36,10 +36,9 @@ date: 31 August 2024
 bibliography: paper.bib
 
 ---
-
 # Summary
 
-`AniSOAP` is a package that translates coarse-grained molecular configurations into tensorial representations that are ideal for supervised machine-learning models of thermodynamic quantities and unsupervised data-driven analyses. This generalization of existing spherical ML representations therefore aims to bridge the gap between two scientific communities: the machine-learned atomistic simulation community, whose primary concern is obtaining fast and (quantum) accurate descriptions of the complex interactions occuring between (spherical) atoms, and the coarse-grained and colloid modeling community, whose primary concern is understanding emergent behavior of macroscopic particles with (plausibly) complex geometries. `AniSOAP` provides a common framework to answer scientific questions at the intersection of these two fields.
+`AniSOAP` is a package that translates coarse-grained molecular configurations into tensorial representations that are ideal for supervised machine-learning models of thermodynamic quantities and unsupervised data-driven analyses. It generalizes several popular spherical representations for atomistic ML and aims to bridge the gap between two scientific communities: the machine-learning enabled atomistic simulation community, which leverages ML to accelerate and refine quantum modeling of complex interactions between spherical atoms, and the coarse-grained and colloid modeling community, which focuses on understanding emergent behavior of macroscopic particles with (plausibly) complex geometries. AniSOAP provides a common framework to tackle scientific questions at the intersection of these two fields.
 
 # Statement of need
 
