### 2025-06-23 #7954 @arthur-lin1027
Hi all. Thank you for your patience as we hopefully near the end of the review process. Just as an update, we are just about done revamping the paper and the documentation over at this PR: https://github.com/cersonsky-lab/AniSOAP/pull/78. The paper has more implementation details, and is followed by more extensive examples in our documentation. I hope this addresses @SamTov's concerns, and we are aiming to get this PR merged into main early this week. I will notify you all here when this happens.
