### 2025-06-27 b2e8bf7
make mic a kwarg
