### 2025-06-30 74c84e7
added mic=mic to get_quat func
