### 2025-07-03 78c831d
Added DOIs for NICE, librascal software, and Dusson22
