+++ paper/paper.bib
@@ -45,6 +45,7 @@ series = {NIPS '22}
 @article{dusson2022,
 title={Atomic cluster expansion: Completeness, efficiency and stability},
 author={Dusson, Genevieve and Bachmayr, Markus and Cs{'a}nyi, G{'a}bor and Drautz, Ralf and Etter, Simon and van Der Oord, Cas and Ortner, Christoph},
+DOI = {10.1016/j.jcp.2022.110946},
 journal={Journal of Computational Physics},
 volume={454},
 pages={110946},
@@ -55,6 +56,7 @@ publisher={Elsevier}
 @article{nice,
 title={Recursive evaluation and iterative contraction of N-body equivariant features},
 author={Nigam, Jigyasa and Pozdnyakov, Sergey and Ceriotti, Michele},
+DOI = {10.1063/5.0021116},
 journal={The Journal of chemical physics},
 volume={153},
 number={12},
@@ -85,6 +87,7 @@ Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan,
 @software{librascal,
   title = {librascal},
   year = {2021},
+  DOI = {10.5281/zenodo.4526062},
   url = {https://github.com/lab-cosmo/librascal}
 }
 
