### 2025-07-03 93c5cef
removed librascal software citation from paper and quippy from bibtex (not used)
