+++ paper/paper.bib
@@ -84,19 +84,6 @@ Guillaume Fraux, Philip Loche, Sergei Kliavinek, Kevin Kazuki Huguenin-Dumittan,
   year = {2025},
   url = {https://github.com/metatensor/featomic}
 }
-@software{librascal,
-  title = {librascal},
-  year = {2021},
-  DOI = {10.5281/zenodo.4526062},
-  url = {https://github.com/lab-cosmo/librascal}
-}
-
-@software{quippy,
-  title={QUIP and quippy documentation},
-  author={Bernstein, Noam and Csányi, Gábor and Kermode, James},
-  year={2020},
-  url={https://libatoms.github.io/QUIP/}
-}
 
 @article{dscribe,
   author = {Himanen, Lauri and J{\"a}ger, Marc O.~J. and Morooka, Eiaki V. and Federici Canova, Filippo and Ranawat, Yashasvi S. and Gao, David Z. and Rinke, Patrick and Foster, Adam S.},
