### 2025-07-03 4f29f6d
removed librascal software citation from paper and quippy from bibtex (not used)
