### 2025-07-03 a509a3a
Revert "removed librascal software citation from paper and quippy from bibtex (not used)"
This reverts commit 4f29f6d18f4b2b43440952643560cf6ab9875816.
