### 2025-07-03 e5f1ed5
deleted the quippy citation, unused
