### 2025-07-03 #7954 @arthur-lin1027
Hi @mbobra,

Thank you for your review! We've updated the DOIs and have made a zenodo here: https://zenodo.org/records/15800936

Is there anything left on the checklist?
