### 2025-07-04 #7954 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
