### 2025-07-08 #7954 @crvernon
@arthur-lin1027 and @rosecers - I just need you to address the following before I accept this submission for publication:

In your archive:

- [ ] The license declared in your Zenodo archive should match the one you specify in your repository.  You can simply edit the metadata to change this.  No new release required.
- [ ] The authors listed and their order should match what is in your paper exactly.  Please edit these to match.

- [ ]  LINE 23:  Missing space between "...explicitly" and "(Bartók et ..."
- [ ]  LINE 24:  Missing space between "...implicitly" and "(Batatia et ..."
- [ ] Actually, this error occurs throughout the paper.  Please correct each of these missing space issues between text and in-text citations.
- [ ] LINES 59-60:  "Specific use-cases of AniSOAP can be seen in our paper here (Lin et al., 2024)." should be edited to read "Lin et al. (2024) provided demonstrations of specific use-cases of AniSOAP."
- [ ] LINE 71:  The "g" in "gaussian" should be capitalized.
- [ ] LINE 71:  "...Gaussian, as outlined in (Lin et al., 2024)." should be written as "...Gausian (Lin et al., 2024)."
- [ ] LINE 82:  "numpy" should be written "NumPy"
- [ ] LINE 93:  Acronym "CI" is used before it is described as "Continuous Integration (CI)"

Let me know when these changes have been made.  Thanks.
