### 2025-07-08 b4b13e3
fixed everything except the Lin et. al citation -- need to look at JOSS formatting specs
