### 2025-07-09 ddd0869
fixed the lin et al citation
