+++ paper/paper.md
@@ -49,7 +49,7 @@ While many implementations construct spherical atomistic descriptors (e.g. DScri
 This is especially relevant as many systems cannot be simulated with all-atom resolution in reasonable times, despite high-performance capabilities of machine-learning interatomic potentials. Additionally, from a conceptual standpoint, we may not always want to analyze the behavior of _atoms_, but superatomic entities, such as functional groups. 
 Most CG techniques still reduce macromolecules to a set of spherical beads. While often adequate for dilute simulations, this reduction to spherical beads significantly oversimplifies anisotropy, which is curcial in condensed systems (e.g., liquid crystals, glasses, molecular crystals). The flexibility of the AniSOAP representation coupled with learning algorithms can help address these challenges, as we can infer anisotropic coarse-graining from high-quality, first-principles data. This featurization could be easily plugged into other software, enabling mesoscopic simulations and enable data-driven insights to many chemical systems at different time and length scales.
 
-The `AniSOAP` package enables the creation of AniSOAP feature vectors, which represent systems of ellipsoidal particles. Analogous to how SOAP or ACE create _atom_-centered representations, `AniSOAP` creates _particle_-centered representations, where a particle could be a single atom or a coarsened group of several atoms. Specific use-cases of AniSOAP can be seen in our paper here [@lin_expanding_2024].
+The `AniSOAP` package enables the creation of AniSOAP feature vectors, which represent systems of ellipsoidal particles. Analogous to how SOAP or ACE create _atom_-centered representations, `AniSOAP` creates _particle_-centered representations, where a particle could be a single atom or a coarsened group of several atoms. @lin_expanding_2024 provided demonstrations of specific use-cases of AniSOAP.
 
 # Implementation details
 The AniSOAP package currently takes in as input a list of frames in the `Atomic Simulation Environment` package [@hjorth_larsen_atomic_2017]. Each frame contains the particles' positions, dimensions, and orientations. If using periodic boundary conditions, the frame also needs to contain the dimensions and orientations of the unit cell. Additional information about each frame can also be stored (e.g. the system energy) and used as a target for supervised ML.
