### 2025-07-09 4d5e499
Last few edits for Chris (#82)
* fixed everything except the Lin et. al citation -- need to look at JOSS formatting specs

* fixed the lin et al citation
