### 2025-07-10 #7954 @crvernon
🥳 Congratulations on your new publication @rosecers and @arthur-lin1027 ! Many thanks to @mbobra for editing and @SamTov and @DaniBodorfor your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
